+++ pkg/shell/grype.go
@@ -1,6 +1,7 @@
 package shell
 
 import (
+	"fmt"
 	"io"
 )
 
@@ -13,6 +14,23 @@ type grypeCmd struct {
 // shell: `grype version`
 func (g *grypeCmd) Version() *Command {
 	return NewCommand(g.InitCmd().WithArgs("version"))
+}
+
+// ScanSBOM runs a grype scan against a target sbom file and produces a JSON report
+//
+// shell: `grype sbom:<filename> --add-cpes-if-none --by-cve -o json
+func (g *grypeCmd) ScanSBOM(filename string) *Command {
+	// TODO: At some point it would be more efficient to read the file once rather than saving to disk
+
+	exe := g.InitCmd().WithArgs(
+		fmt.Sprintf("sbom:%s", filename),
+		"--add-cpes-if-none",
+		"--by-cve",
+		"-o",
+		"json",
+	)
+
+	return NewCommand(exe)
 }
 
 // GrypeCommand with custom stdout and stderr